backend/requirements.txt
Django==5.2.5
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
channels==4.0.0
Pillow==10.1.0 